refactor(admin): trocea el god-service en 5 servicios de dominio
COD-02: admin.service.ts (~1489 lineas, ~10 dominios) eliminado y dividido
en AdminUsers/Content/Analytics/Gamification/Exams Service, cada uno con sus
deps. Controller inyecta los 5 y enruta los 39 endpoints; comportamiento
identico. Los 21 tests de usuarios migran a admin-users.service.spec.ts.
