+++ apps/api/src/admin/admin.module.ts
@@ -1,6 +1,10 @@
 import { Module } from '@nestjs/common';
 import { AdminController } from './admin.controller';
-import { AdminService } from './admin.service';
+import { AdminUsersService } from './admin-users.service';
+import { AdminContentService } from './admin-content.service';
+import { AdminAnalyticsService } from './admin-analytics.service';
+import { AdminGamificationService } from './admin-gamification.service';
+import { AdminExamsService } from './admin-exams.service';
 import { CourseGeneratorService } from './course-generator.service';
 import { BillingService } from './billing.service';
 import { CertificatesModule } from '../certificates/certificates.module';
@@ -9,6 +13,14 @@ import { YoutubeModule } from '../youtube/youtube.module';
 @Module({
   imports: [CertificatesModule, YoutubeModule],
   controllers: [AdminController],
-  providers: [AdminService, CourseGeneratorService, BillingService],
+  providers: [
+    AdminUsersService,
+    AdminContentService,
+    AdminAnalyticsService,
+    AdminGamificationService,
+    AdminExamsService,
+    CourseGeneratorService,
+    BillingService,
+  ],
 })
 export class AdminModule {}
